api/requirements.txt
anyio==3.6.1
fastapi==0.79.0
idna==3.3
psycopg2-binary==2.9.3
pydantic==1.9.1
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2